Organized files into a new directory structure.
