+++ src/vehicle/simulation/SimulatedVehicle.cpp
