+++ src/vehicle/testplatform/Gpio.cpp
