+++ src/vehicle/testplatform/HardwareIO.h
