+++ src/vehicle/testplatform/HardwareIOFile.cpp
