+++ src/vehicle/testplatform/HardwareIOFile.h
