+++ src/vehicle/testplatform/Motor.cpp
