+++ src/vehicle/testplatform/Motor.h
