+++ src/vehicle/testplatform/MotorConstants.h
