+++ src/vehicle/testplatform/StateTracker.cpp
