+++ src/vehicle/testplatform/SystemGPIOs.cpp
