+++ src/vehicle/testplatform/SystemGPIOs.h
