+++ src/vehicle/testplatform/TestPlatform.cpp
